♻️ [Refactoring]: #455 Phase 8 unknown_escape カテゴリを tests/unknown_escape.rs に分離
未知エスケープ（バックスラッシュ脱落）テスト 3 件を
literal_string/tests/unknown_escape.rs に切り出し、tests.rs ハブに
`mod unknown_escape;` 宣言を追加。
